fix(ext/napi): ensure the finalizer callback will be called (#29710)
See
https://github.com/napi-rs/napi-rs/issues/2708#issuecomment-2963957944
for the context.

In current implementation, the Weak `Reference` is created with
`v8::Weak::with_finalizer`, but there is no guarantee as to *when* or
even *if* the finalization callback:
https://github.com/denoland/rusty_v8/blob/v137.2.0/src/handle.rs#L623-L627.
It may cause the memory leak if Node-API caller want to do some cleanup
jobs in the finalization callback.