perf: Use more efficient structuredClone API (#30258)
This PR builds on top of https://github.com/denoland/deno_core/pull/1167
that provided 16% perf improvement for structured clone algorithm by avoid round-trip
serialization and deserialization.

This was applied to the internal `structuredClone` helper, as well as
for `MessagePort.postMessage` which is what Web worker 
and `node:worker_thread` APIs are using.