fix(install): don't try to re-symlink binaries in node_modules if the symlink is correct (#30546)
Fix #30424.

I think this should avoid the errors. It will add some overhead when
they aren't set up yet, but binaries are pretty small in number so I
doubt it will matter.

In the case where it's already set up (which is probably common), it
should avoid some work and prevent using binaries that are already in
use