perf: speed up file diffing
This swaps out dissimilar for imara which is substantially faster at diffing strings.

Note that this is a proof of concept and I did not have enough time to make the output pretty. I just shows that the diff is fast. Applying colors should be doable without changing much about the perf.

Fixes https://github.com/denoland/deno/issues/30634