feat(unstable): support `Deno.HttpClient` in `WebSocket` (#30692)
This commit adds support for specifying a `Deno.HttpClient` in the
`client` field of the `WebSocketOptions` bag. This mirrors how users
pass a custom HTTP client to `fetch`.

Additionally the underlying connection for websockets are now
established through the same mechanism used for `fetch`, which
means that `WebSocket` now correctly uses HTTP proxies, like
when configuring the `HTTP_PROXY` env var.

---------

Signed-off-by: Luca Casonato <hello@lcas.dev>