fix(ext/node): `fs.stat` and `fs.statSync` compatibility (#30637)
Closes #30570

Changes in this PR:
- Implement `ino`, `nlink`, and `blocks` properties of `Deno.FileInfo`
on Windows. These changes are automatically reflected to the
corresponding node stat function. In order to do so, I had to tinker
with the
[createByteStruct](https://github.com/denoland/deno/blob/a3a904da1450849b51c34a972db4fd9b9a572d29/ext/fs/30_fs.js#L297)
function to create another optional int type, apart from `?u64`. It's
common for small sized files on Windows (particularly NTFS file system)
to have a `Stats.blocks` property of 0, and currently all 0 values with
type `?u64` will be coerced into `null` by `createByteStruct`.
- Refactor the `BigIntStats` and `Stats` class, to use the same class
with Node.js that are provided from
[utils.mjs](https://github.com/denoland/deno/blob/7f8e488c3636e76f15bc1b33693e70b2ebf0593d/ext/node/polyfills/internal/fs/utils.mjs#L577).
Also ensures that all properties are not `null` or `undefined`.
- Addresses the `prefer-primordials` lint rule.