fix(webgpu): crash after isolate disposal (#30909)
this loop doesn't inherently stop running when the isolate stops
running, so use the isolate task spawner to ensure logic only runs when
the event loop is still running. it would be nice to move this logic
inside the DeviceErrorHandler (we could get rid of the task loop!) but
it creates a cyclic dependency on the device object which was annoying
to think about.