feat(internal): support variants of spec tests (#30947)
Adds support for a new key `"variants"` that effectively parameterizes a
test, allowing you to substitute different values for each scenario. For
instance, I'm planning on using this in the tsgo PR, where I'll have
tests like

```json
{
  "variants": {
    "tsgo": {
      "maybe_unstable_tsgo": "--unstable-tsgo"
    },
    "tsjs": {
      "maybe_unstable_tsgo": ""
    }
  },
  "steps": [
    {
      "args": "install",
      "output": "[WILDCARD]"
    },
    {
      "args": "check ${maybe_unstable_tsgo} main.ts",
      "output": "check.out"
    }
  ]
}
```
This effectively expands into two separate tests: one called
`test_name::tsgo` which runs `check --unstable-tsgo main.ts`, and
another called `test_name::tsjs` which runs `check main.ts`

I imagine this to be useful to run tests under the various
`--node-modules` modes as well

The way it works is that it generates a test for each variant, and then
in each variant the values specified for each variant are substituted
whenever you refer to them with `${name}` (I added the curly braces to
make it unambiguous when you have variables that are prefixes of one
another). Right now that substitution only applies to the `"output"` key
(so you can have different output expectations for different variants)
and `"args"`, but we can easily expand that

Implementation is not particularly efficient, but it's good enough for
now