refactor: merge `deno_url` into `deno_web` (#31182)
This commit deprecated `deno_url` crate and merges it into `deno_web`.

This will allow us to limit number of crates we need to publish and
(maybe) improve compile and link times.

The actual `ext/url` directory will be removed in a follow up PR, once a
new version is published and points to `deno_web` crate.