refactor: merge `deno_cache` into `deno_web` (#31183)
This commit deprecated `deno_cache` crate and merges it into `deno_web`.

This will allow us to limit number of crates we need to publish and
(maybe) improve compile and link times.

The actual `ext/cache` directory will be removed in a follow up PR, once
a new version is published and points to `deno_web` crate.

---------

Co-authored-by: Bartek Iwańczuk <biwanczuk@gmail.com>