refactor: merge `deno_console` into `deno_web` (#31185)
This commit deprecated `deno_console` crate and merges it into
`deno_web`.

This will allow us to limit number of crates we need to publish and
(maybe) improve compile and link times.

The actual `ext/console` directory will be removed in a follow up PR,
once a new version is published and points to `deno_web` crate.

---------

Co-authored-by: Bartek Iwańczuk <biwanczuk@gmail.com>