fix(lsp): `organizeImports` without resolving specifiers
Organize imports action should move, group, or delete imports but not
rewrite the imported modules.

The current behavior is causing a bug with workspaces, rewriting
relative imports to scoped imports if the workspace members are loaded
in the server. The two specifiers are not functionally equivalent, at
least due to https://github.com/jsr-io/jsr/issues/1220.

This change removes the rewriting and keeps the imports as they are sent
from tsc.

Fixes #31141

[ci]