fix: ensure that the node:console implementation has an implementation for `emitWarning` in scope
Introduces the emitWarning function from the process module to enable warning capabilities within the console constructor implementation.

This fixes a bug where the `time`, `countReset` and `timeLogImpl` functions would throw an error due to trying to call `emitWarning`, they would throw an error for an undefined variable named `emitWarning` or if a global function had that name` they would call that function, which is not expected behaviour here.