fix(ext/node): move sqlite_extension into separate module from sqlite_extension_test to permit loadable_extension feature
This allows the rusqlite `loadable_extension` feature to be used -- without it, we were building an extension that triggered a segfault on load (because it wasn't using the pointer provided during extension load to access the previously-initialized sqlite library instance). Because that feature conflicts with other features used _outside_ the `sqlite_extension_test` module, we split off to a separate workspace.

The test harness stays in `sqlite_extension_test`, as it depends on other parts of the workspace.