fix(tsgo): do not error out on non js/ts imports from npm packages (#31478)
Fixes #31423 

This already worked for non npm files because the loading for non-npm
files uses the module graph, and things like CSS don't appear in the
module graph. That meant we weren't prompting tsc to create a source
file and parse those files. For npm files though, we were loading them
manually, which meant tsc tried to load them as ts/js sources