fix(ext/node): fix sqlite extension used for testing; ensure related tests are actually meaningful (#31455)
The SQLite extension-loading tests, prior to this PR, were
short-circuiting due to an incorrect search path, and not actually
verifying that extensions loaded. When this invalid path was fixed, the
result was a segfault due to an error in the extension itself (failing
to use the pointer provided at initialization time to access the
existing sqlite instance).

This is addressed in three different stages:

1. Switch to using `ignore` rather than an early return when tests
cannot be run due to the extension not existing. *Running
`sqlite_extension_test.ts` in this state allows the original problem, of
tests not actually being meaningfully run, to be observed.*
2. Fix the location used to run sqlite extension tests. *This fixes the
problem exposed by the above step, but causes a segmentation fault at
extension load time.*
3. Split extension build into a separate module, written to use
rusqlite's `loadable_extension` support. *This fixes the segfault
exposed above.*

Because the sysroot used for x86_64 Linux builds does not provide
sqlite, and rusqlite's extension build mechanism attempts to link
against libsqlite3.so regardless of whether the `bundled` flag is used,
this PR does not use the sysroot for the extension build.

Fixes #31454