fix(ext/telemetry): otel baggage propagation (#31620)
There were three bugs in baggage propagation:

- `BaggageImpl` was not reading items from the map passsed in, instead
always creating an empty baggage
- `Deno.serve` was reading out baggage and storing it in a context, but
then not starting the callback in that context, but in the parent
context, meaning the user did not have access to the baggage.
- The baggage was stored on the context with the wrong key.

The gist of it is that `propagation.getActiveBaggage()` works now.

Secondly, this adds a composite propagator into the global OTEL context,
so that `propagation.extract()` and `propagation.inject()` start
working.

Fixes #28085 and fixes #30064, towards #28082